+++ CHANGELOG.md
@@ -20,6 +20,7 @@ All notable changes to this project will be documented in this file.
 - Fix REPL view input indentation
 - Performance improvements
 - Fix caret position(s) after backward slurp
+- Add Forward/Backward Up/Down ParEdit commands #9
 
 ## 0.6.0 - 2020-11-20
 - Add the **Tutkain: Evaluate** command. Example: